[pentest] Filter state & counter for lc_lctrl_fi test
Depending on the SKU config, the target could be in a different life
cycle state and life cycle counter. Hence, it does not make sense to
compare these two values with the testvector. However, if there would
be a mismatch between LC state & counter before and after the FI
trigger window, the "res" field indicates this.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>