[pentest] Make tests resilient
Add the flaky_expected_result option where wrong outputs do not call a break of the test but outputs are still logged.
Add the option to reset between tests.
Make changes to aes_sca and ibex_fi to still provide a boolean result as output json when errors are found.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>